+++ aksw-commons-rx/src/main/java/org/aksw/commons/rx/cache/range/SliceWithPagesImpl.java
@@ -81,7 +81,7 @@ public class SliceWithPagesImpl<T>
 
 
     @Override
-    public int getPageSize() {
+    public long getPageSize() {
         return pageSize;
     }
