+++ aksw-commons-util/src/main/java/org/aksw/commons/util/range/PageHelper.java
@@ -1,25 +1,25 @@
 package org.aksw.commons.util.range;
 
 public interface PageHelper {
-    int getPageSize();
+    long getPageSize();
 
     default long getPageIdForOffset(long offset) {
-        int pageSize = getPageSize();
+        long pageSize = getPageSize();
         return getPageIdForOffset(offset, pageSize);
     }
 
-    default int getIndexInPageForOffset(long offset) {
-        int pageSize = getPageSize();
+    default long getIndexInPageForOffset(long offset) {
+        long pageSize = getPageSize();
         return getIndexInPageForOffset(offset, pageSize);
     }
 
-    public static long getPageIdForOffset(long offset, int pageSize) {
+    public static long getPageIdForOffset(long offset, long pageSize) {
         long result = offset / pageSize;
         return result;
     }
 
-    public static int getIndexInPageForOffset(long offset, int pageSize) {
-        return (int)(offset % (long)pageSize);
+    public static long getIndexInPageForOffset(long offset, long pageSize) {
+        return offset % pageSize;
     }
 
 }
